+++ ClimateExplorer.Core/Model/DataFileFilterAndAdjustment.cs
@@ -4,5 +4,5 @@ public class DataFileFilterAndAdjustment
     public required string Id { get; set; }
     public DateTime? StartDate { get; set; }
     public DateTime? EndDate { get; set; }
-    public float? ValueAdjustment { get; set; }
+    public double? ValueAdjustment { get; set; }
 }
